data/dags/requirements.txt
python-dotenv
psycopg2-binary
google-cloud-bigquery